+++ bundles/org.eclipse.passage.lic.api/src/org/eclipse/passage/lic/api/version/SemanticVersion.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,12 +10,14 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.api.version;
+package org.eclipse.passage.lic.api.version;
 
 /**
  * <p>
  * Well structured representation of a version.
  * </p>
+ * 
+ * @since 1.1
  */
 public interface SemanticVersion {
