+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/version/NumericalVersion.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,13 +10,13 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.version;
+package org.eclipse.passage.lic.base.version;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.Supplier;
 
-import org.eclipse.passage.lic.internal.api.version.SemanticVersion;
+import org.eclipse.passage.lic.api.version.SemanticVersion;
 
 /**
  * Auxiliary unit operating over {@linkplain SemanticVersion} collects it's
