+++ bundles/org.eclipse.passage.lic.equinox/src/org/eclipse/passage/lic/equinox/requirements/RequirementFromComponent.java
@@ -19,10 +19,10 @@ import java.util.function.Supplier;
 import org.eclipse.passage.lic.api.requirements.Requirement;
 import org.eclipse.passage.lic.base.requirements.BaseFeature;
 import org.eclipse.passage.lic.base.requirements.BaseRequirement;
+import org.eclipse.passage.lic.base.version.DefaultVersion;
+import org.eclipse.passage.lic.base.version.SafeVersion;
 import org.eclipse.passage.lic.internal.api.restrictions.RestrictionLevel;
 import org.eclipse.passage.lic.internal.base.restrictions.DefaultRestrictionLevel;
-import org.eclipse.passage.lic.internal.base.version.DefaultVersion;
-import org.eclipse.passage.lic.internal.base.version.SafeVersion;
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.runtime.dto.ComponentDescriptionDTO;
