+++ bundles/org.eclipse.passage.lic.equinox/src/org/eclipse/passage/lic/internal/equinox/ApplicationVersion.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -17,9 +17,9 @@ import java.util.function.Supplier;
 
 import org.eclipse.equinox.app.IApplicationContext;
 import org.eclipse.passage.lic.base.ProductVersion;
-import org.eclipse.passage.lic.internal.base.version.BaseSemanticVersion;
-import org.eclipse.passage.lic.internal.base.version.DefaultVersion;
-import org.eclipse.passage.lic.internal.base.version.SafeVersion;
+import org.eclipse.passage.lic.base.version.BaseSemanticVersion;
+import org.eclipse.passage.lic.base.version.DefaultVersion;
+import org.eclipse.passage.lic.base.version.SafeVersion;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.Version;
