+++ tests/org.eclipse.passage.lic.api.tests/src/org/eclipse/passage/lic/api/tests/version/SemanticVersionContractTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -16,10 +16,9 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assume.assumeFalse;
 import static org.junit.Assume.assumeTrue;
 
-import org.eclipse.passage.lic.internal.api.version.SemanticVersion;
+import org.eclipse.passage.lic.api.version.SemanticVersion;
 import org.junit.Test;
 
-@SuppressWarnings("restriction")
 public abstract class SemanticVersionContractTest {
 
 	@Test(expected = RuntimeException.class)
