+++ tests/org.eclipse.passage.lic.base.tests/src/org/eclipse/passage/lic/internal/base/tests/version/contract/SafeVersionContractTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -13,8 +13,8 @@
 package org.eclipse.passage.lic.internal.base.tests.version.contract;
 
 import org.eclipse.passage.lic.api.tests.version.VersionContractTest;
-import org.eclipse.passage.lic.internal.api.version.Version;
-import org.eclipse.passage.lic.internal.base.version.SafeVersion;
+import org.eclipse.passage.lic.api.version.Version;
+import org.eclipse.passage.lic.base.version.SafeVersion;
 
 @SuppressWarnings("restriction")
 public final class SafeVersionContractTest extends VersionContractTest {
