+++ tests/org.eclipse.passage.lic.equinox.tests/OSGI-INF/l10n/bundle.properties
@@ -1,6 +1,6 @@
 #Properties file for org.eclipse.passage.lic.equinox.tests
 ###############################################################################
-# Copyright (c) 2018, 2020 ArSysOp and others
+# Copyright (c) 2018, 2021 ArSysOp and others
 #
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
